Resolved conflicts in modules/apt.py.

+++ salt/modules/timezone.py
@@ -1,25 +1,23 @@
 '''
-Module for managing timezone on posix-like systems.
+Module for managing timezone on POSIX-like systems.
 '''
 
 # Import python libs
 import os
 import hashlib
-import salt.utils
 import logging
+
+# Import salt libs
+import salt.utils
 
 log = logging.getLogger(__name__)
 
 
 def __virtual__():
     '''
-    Only work on posix-like systems
+    Only work on POSIX-like systems
     '''
-    # Disable on these platorms, specific service modules exist:
-    disable = [
-        'Windows',
-        ]
-    if __grains__['os'] in disable:
+    if salt.utils.is_windows():
         return False
     return 'timezone'
 
@@ -34,18 +32,23 @@ def get_zone():
     '''
     cmd = ''
     if 'Arch' in __grains__['os_family']:
-        cmd = 'grep TIMEZONE /etc/rc.conf | grep -vE "^#"'
+        cmd = ('timedatectl | grep Timezone |'
+               'sed -e"s/: /=/" -e"s/^[ \t]*//" | cut -d" " -f1')
     elif 'RedHat' in __grains__['os_family']:
         cmd = 'grep ZONE /etc/sysconfig/clock | grep -vE "^#"'
     elif 'Suse' in __grains__['os_family']:
         cmd = 'grep ZONE /etc/sysconfig/clock | grep -vE "^#"'
     elif 'Debian' in __grains__['os_family']:
-        return open('/etc/timezone','r').read()
+        with salt.utils.fopen('/etc/timezone', 'r') as ofh:
+            return ofh.read()
     elif 'Gentoo' in __grains__['os_family']:
-        return open('/etc/timezone','r').read()
+        with salt.utils.fopen('/etc/timezone', 'r') as ofh:
+            return ofh.read()
     elif 'FreeBSD' in __grains__['os_family']:
         return ('FreeBSD does not store a human-readable timezone. Please'
                 'consider using timezone.get_zonecode or timezone.zonecompare')
+    elif 'Solaris' in __grains__['os_family']:
+        cmd = 'grep "TZ=" /etc/TIMEZONE'
     out = __salt__['cmd.run'](cmd).split('=')
     ret = out[1].replace('"', '')
     return ret
@@ -89,24 +92,37 @@ def set_zone(timezone):
 
         salt '*' timezone.set_zone 'America/Denver'
     '''
-    zonepath = '/usr/share/zoneinfo/{0}'.format(timezone)
+    if 'Solaris' in __grains__['os_family']:
+        zonepath = '/usr/share/lib/zoneinfo/{0}'.format(timezone)
+    else:
+        zonepath = '/usr/share/zoneinfo/{0}'.format(timezone)
     if not os.path.exists(zonepath):
         return 'Zone does not exist: {0}'.format(zonepath)
 
     if os.path.exists('/etc/localtime'):
         os.unlink('/etc/localtime')
-    os.symlink(zonepath, '/etc/localtime')
+
+    if 'Solaris' in __grains__['os_family']:
+        __salt__['file.sed'](
+            '/etc/default/init', '^TZ=.*', 'TZ={0}'.format(timezone))
+    else:
+        os.symlink(zonepath, '/etc/localtime')
 
     if 'Arch' in __grains__['os_family']:
-        __salt__['file.sed']('/etc/rc.conf', '^TIMEZONE=.*', 'TIMEZONE="{0}"'.format(timezone))
+        __salt__['file.sed'](
+            '/etc/rc.conf', '^TIMEZONE=.*', 'TIMEZONE="{0}"'.format(timezone))
     elif 'RedHat' in __grains__['os_family']:
-        __salt__['file.sed']('/etc/sysconfig/clock', '^ZONE=.*', 'ZONE="{0}"'.format(timezone))
+        __salt__['file.sed'](
+            '/etc/sysconfig/clock', '^ZONE=.*', 'ZONE="{0}"'.format(timezone))
     elif 'Suse' in __grains__['os_family']:
-        __salt__['file.sed']('/etc/sysconfig/clock', '^ZONE=.*', 'ZONE="{0}"'.format(timezone))
+        __salt__['file.sed'](
+            '/etc/sysconfig/clock', '^ZONE=.*', 'ZONE="{0}"'.format(timezone))
     elif 'Debian' in __grains__['os_family']:
-        open('/etc/timezone', 'w').write(timezone)
+        with salt.utils.fopen('/etc/timezone', 'w') as ofh:
+            ofh.write(timezone)
     elif 'Gentoo' in __grains__['os_family']:
-        open('/etc/timezone', 'w').write(timezone)
+        with salt.utils.fopen('/etc/timezone', 'w') as ofh:
+            ofh.write(timezone)
 
     return True
 
@@ -121,15 +137,19 @@ def zone_compare(timezone):
 
         salt '*' timezone.zone_compare 'America/Denver'
     '''
-    if not os.path.exists('/etc/localtime'):
-        return 'Error: /etc/localtime does not exist.'
+    if 'Solaris' in __grains__['os_family']:
+        return 'Not implemented for Solaris family'
 
+    tzfile = '/etc/localtime'
     zonepath = '/usr/share/zoneinfo/{0}'.format(timezone)
+
+    if not os.path.exists(tzfile):
+        return 'Error: {0} does not exist.'.format(tzfile)
 
     with salt.utils.fopen(zonepath, 'r') as fp_:
         usrzone = hashlib.md5(fp_.read()).hexdigest()
 
-    with salt.utils.fopen('/etc/localtime', 'r') as fp_:
+    with salt.utils.fopen(tzfile, 'r') as fp_:
         etczone = hashlib.md5(fp_.read()).hexdigest()
 
     if usrzone == etczone:
@@ -157,16 +177,31 @@ def get_hwclock():
         cmd = 'tail -n 1 /etc/adjtime'
         return __salt__['cmd.run'](cmd)
     elif 'Debian' in __grains__['os_family']:
+        #Original way to look up hwclock on Debian-based systems
         cmd = 'grep "UTC=" /etc/default/rcS | grep -vE "^#"'
         out = __salt__['cmd.run'](cmd).split('=')
-        if out[1] == 'yes':
-            return 'UTC'
+        if len(out) > 1:
+            if out[1] == 'yes':
+                return 'UTC'
+            else:
+                return 'localtime'
         else:
-            return 'localtime'
+            #Since Wheezy
+            cmd = 'tail -n 1 /etc/adjtime'
+            return __salt__['cmd.run'](cmd)
     elif 'Gentoo' in __grains__['os_family']:
         cmd = 'grep "^clock=" /etc/conf.d/hwclock | grep -vE "^#"'
         out = __salt__['cmd.run'](cmd).split('=')
         return out[1].replace('"', '')
+    elif 'Solaris' in __grains__['os_family']:
+        if os.path.isfile('/etc/rtc_config'):
+            with salt.utils.fopen('/etc/rtc_config', 'r') as fp_:
+                for line in fp_:
+                    if line.startswith('zone_info=GMT'):
+                        return 'UTC'
+            return 'localtime'
+        else:
+            return 'UTC'
 
 
 def set_hwclock(clock):
@@ -178,26 +213,45 @@ def set_hwclock(clock):
         salt '*' timezone.set_hwclock UTC
     '''
     timezone = get_zone()
-    zonepath = '/usr/share/zoneinfo/{0}'.format(timezone)
+
+    if 'Solaris' in __grains__['os_family']:
+        if 'sparc' in __grains__['cpuarch']:
+            return 'UTC is the only choice for SPARC architecture'
+        if clock == 'localtime':
+            cmd = 'rtc -z {0}'.format(timezone)
+            __salt__['cmd.run'](cmd)
+            return True
+        elif clock == 'UTC':
+            cmd = 'rtc -z GMT'
+            __salt__['cmd.run'](cmd)
+            return True
+    else:
+        zonepath = '/usr/share/zoneinfo/{0}'.format(timezone)
+
     if not os.path.exists(zonepath):
         return 'Zone does not exist: {0}'.format(zonepath)
 
-    os.unlink('/etc/localtime')
-    os.symlink(zonepath, '/etc/localtime')
+    if 'Solaris' not in __grains__['os_family']:
+        os.unlink('/etc/localtime')
+        os.symlink(zonepath, '/etc/localtime')
 
     if 'Arch' in __grains__['os_family']:
-        __salt__['file.sed']('/etc/rc.conf', '^HARDWARECLOCK=.*', 'HARDWARECLOCK="{0}"'.format(clock))
+        __salt__['file.sed'](
+            '/etc/rc.conf', '^HARDWARECLOCK=.*', 'HARDWARECLOCK="{0}"'.format(
+                clock))
     elif 'RedHat' in __grains__['os_family']:
-        __salt__['file.sed']('/etc/sysconfig/clock', '^ZONE=.*', 'ZONE="{0}"'.format(timezone))
+        __salt__['file.sed'](
+            '/etc/sysconfig/clock', '^ZONE=.*', 'ZONE="{0}"'.format(timezone))
     elif 'Suse' in __grains__['os_family']:
-        __salt__['file.sed']('/etc/sysconfig/clock', '^ZONE=.*', 'ZONE="{0}"'.format(timezone))
+        __salt__['file.sed'](
+            '/etc/sysconfig/clock', '^ZONE=.*', 'ZONE="{0}"'.format(timezone))
     elif 'Debian' in __grains__['os_family']:
         if clock == 'UTC':
             __salt__['file.sed']('/etc/default/rcS', '^UTC=.*', 'UTC=yes')
         elif clock == 'localtime':
             __salt__['file.sed']('/etc/default/rcS', '^UTC=.*', 'UTC=no')
     elif 'Gentoo' in __grains__['os_family']:
-        __salt__['file.sed']('/etc/conf.d/hwclock', '^clock=.*', 'clock="{0}"'.format(clock))
+        __salt__['file.sed'](
+            '/etc/conf.d/hwclock', '^clock=.*', 'clock="{0}"'.format(clock))
 
     return True
-
